[QA | Docs] Add docs & tests for add, refactored is_less (#54)
* [Ref] Refactored less to be bit-by-bit

- This refactoring was needed to make is_less compatible with ints.
Previous version with uints value comparison is not working correctly
with signed ints.
- Also I returned normalization of exponent to less, which fixed
failing tests.

* [Docs] Add notes, fix todo, found bug in sub

+++ src/tests/helpers/int128_converters.c
@@ -54,7 +54,7 @@ long long get_random_ll(void) {
 
 short int get_random_short(void) {
     short int val = 0;
-    for (uint64_t i = 0; i < 16; i++)
+    for (int i = 0; i < 16; i++)
         if (rand() % 2)
             ADD_BIT(val, i);
